Fix audio+video sync and frame rate issues.
This version ignores any initial audio or video delay values (I'm
assuming both streams *should* start at zero. Using this method seems to
solve issues with "non monotone timestamps" as well as fixes some of the
sync problems we've seen.

The main trick about the frame rate is that newer ffmpeg versions set a
ticks_per_frame multiplier (of 2) on the time base, but this isn't
accounted for in the segmenter. So I added something to take that into
account.

For those having trouble with the video playing twice as fast as it
should and the audio being off, these are the changes I had to make to
get it working.